+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This file is a history of the changes made to @idearium/cli.
 
-## Unreleased
+## v6.1.0-beta.1 - 2026-09-15
 
 ### Added